ProyectoPythonParkingFlask/requirements.txt
Flask
Flask-Script
Flask-Bootstrap4
Flask-Sqlalchemy
python-dateutil
